% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsgif.R
\name{get_wsgif}
\alias{get_wsgif}
\title{get_wsgif}
\usage{
get_wsgif(flow_fn_1, flow_fn_2, flow_file_dir, output_fn_tag = "",
  output_file_dir, historical_period = c("1955-01-01", "2004-12-31"),
  future_period = c("2005-01-01", "2095-12-31"), by = "all",
  dam_cutoff_percentile = 1)
}
\arguments{
\item{flow_fn_1}{file naming format for the hydro_flow_file, excluding the region (e.g., 10_HIST_MOSART.csv, would be "HIST_MOSART.csv")}

\item{flow_fn_2}{optional second flow file naming format in case hist and future parts are split.}

\item{flow_file_dir}{directory where hydro flow files are stored}

\item{output_fn_tag}{naming convention for output files}

\item{output_file_dir}{directory for output. If blank, new directory is made inside hydro_flow_file_dir.}

\item{historical_period}{start and end dates of historical period in format yyyy-mm-dd}

\item{future_period}{start and end dates of future period in format yyyy-mm-dd}

\item{by}{spatial unit ("all", "bal_auth", "region", "HUCWM", "WECC")}

\item{dam_cutoff_percentile}{cutoff for which dams to include in the wsgif computation. The percentile is based on nameplate capacity (i.e., percentile of 0.5 will include the largest dams that contribute 50 percent total nameplate).}
}
\value{
time series of WSGIF
}
\description{
get_wsgif
}
\details{
computes WSGIF for hydropower by balancing authority, region, HUC2, or entire WECC
}
