% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsgif.R
\name{get_wsgif_hydro}
\alias{get_wsgif_hydro}
\title{get_wsgif_hydro}
\usage{
get_wsgif_hydro(hydro_flow_file_1, hydro_flow_file_2,
  historical_period = 1955:2004, future_period = 2005:2095,
  by = "all")
}
\arguments{
\item{hydro_flow_file_1}{directory of results folder}

\item{hydro_flow_file_2}{optional second flow file (in case results are split); will be binded to hydro_flow_file_1}

\item{historical_period}{...}

\item{future_period}{...}

\item{by}{spatial unit ("all", "bal_auth", "region", "HUC2", "WECC")}
}
\value{
time series of WSGIF
}
\description{
get_wsgif_hydro
}
\details{
computes WSGIF for hydropower by balancing authority, region, HUC2, or entire WECC
}
