% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsgif.R
\name{compute_wsgif_hydro}
\alias{compute_wsgif_hydro}
\title{compute_wsgif_hydro}
\usage{
compute_wsgif_hydro(wm_output_dir)
}
\arguments{
\item{wm_output_dir}{directory of results folder}
}
\value{
time series of WSGIF
}
\description{
compute_wsgif_hydro
}
\details{
computes WSGIF for hydropower
}
