% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plexos.R
\name{get_plexos_inputs}
\alias{get_plexos_inputs}
\title{get_plexos_inputs}
\usage{
get_plexos_inputs(scenario_file, flow_dir, huc4_shape_dir)
}
\arguments{
\item{flow_dir}{directory where huc4 flow results reside}

\item{scenario_name}{file naming format for the huc4 flow file}
}
\value{
time series of WSGIF
}
\description{
get_plexos_inputs
}
\details{
derives plexos inputs for given scenario
}
