% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_flow}
\alias{prep_flow}
\title{prep_flow}
\usage{
prep_flow(wm_output_dir, output_dir = "", in_parallel = FALSE)
}
\arguments{
\item{wm_output_dir}{directory of results folder}

\item{in_parallel}{run data extraction in parallel (where multiple regions are being assessed)}

\item{hyd_output_dir}{output directory}
}
\value{
daily time series of regulated flows at hydropower grids
}
\description{
prep_flow
}
\details{
extracts flow data for hydropower grids; combines files if runs are temporally split
}
